[WIP] initial commit for Speech Language Model API support
For SLMs, model is explicitely fed with continuous embedding
as an input at each decoding step. Hidden state is output to
be used in further processing. The `demo_slm_api` notebook
shows the intended use with regular text LLM as an example.

After each output for the given request, the input for the next
iteration is async set with `append_request`.
That calls `core_cli` method and passes the embedding to the backend
through serialization. In `core` the provided embedding is getting
deserialized and passed to `scheduler.set_input_embeds`,
where it is passed to the request.
Background scheduler loop runs `schedule` which checks which
requests have the input being set.
If it is, the input embedding is passed to CachedRequestData
and then to gpu_input_batch. Finally at the execution,
gpu_model_runner puts the embeddings from the input batch to the
buffer that will be passed as input to the model definition.

Currently, the inference always expects the input embedding to be set.
It should be changed to be optional. We should also optionally
disable sampling in case input embeddings are set explicitely.

Signed-off-by: Viacheslav Klimkov <vklimkov@nvidia.com>