[QUICK FIX] avoid non-blocked main loop while waiting for embeddings to be set
Actually processing the messages (request ADD and APPEND), scheduling and executing
inference runs in a single thread in core. When there is no running requests,
the thread is waiting on the input message queue get() method. But if the request
is running, we dont wait for any messages from the queue.
This quick iteration in a non-blocking loop hurts performance.
This quickfix makes sure that if there are no requests that actually have embeddings,
then we should wait for new messages in the input queue.

Signed-off-by: Viacheslav Klimkov <vklimkov@nvidia.com>