[WIP] add inefficient support for bsz>1
The approach for now is to just remove and re-add the ragged tensor
format at the beginning and end of the forward pass. In the future for
performance we may want to add convs that can handle the ragged tensor
format.

The approach also assumes that all seqlens are the same, which is not
yet guaranteed by the scheduler, so we'll have to somehow modify this
later.