Refactor SLM API to accept arbitrary number of custom inputs
Previously it was only possible to pass custom hidden states
to be used during as input to the model. We also had sync execution,
where scheduler waits for the hidden states to be set for the request,
before adding it to the batch. This refactoring generalizes this approach,
allowing to set/append any number of custom inputs to the LLM.

config.json should have custom input specification, such as:
```
"custom_input_specs": [{"name": "some_embs", "dim": 2048}]
```
which tells vLLM that model expects custom input "some_embs"
in the `forward` method. Then in `generate` and `append_request`,
user can provide custom inputs for each decoding step.

Signed-off-by: Viacheslav Klimkov <vklimkov@nvidia.com>